[API-24003] - release notes and sandbox access pages (#1229)
* Base release notes and sandbox access changes

* CSS updates

* Release notes unit tests

* Sandbox success unit tests and CSS tweak

* CSS tweaks

* ReleaseNotes css fixes

* Sandbox css fixes

* Update Flow to Grant in a couple places

* Update pages to use urlSlug where possible

* Unit test fixes and file structure updates
