+++ package.json
@@ -25,7 +25,7 @@
     "@braintree/sanitize-url": "^6.0.2",
     "@department-of-veterans-affairs/component-library": "8.5.0",
     "@department-of-veterans-affairs/formation": "6.16.3",
-    "@department-of-veterans-affairs/sandbox-access-form": "git+ssh://git@github.com:department-of-veterans-affairs/va-api-sandbox-access-form.git#v0.1.1",
+    "@department-of-veterans-affairs/sandbox-access-form": "git+ssh://git@github.com:department-of-veterans-affairs/va-api-sandbox-access-form.git#3c0c20f151ae84941c29a741bd5aff44a2a4fb7b",
     "@department-of-veterans-affairs/web-components": "2.11.0",
     "@floating-ui/react": "^0.14.0",
     "@floating-ui/react-dom": "^1.3.0",
