+++ src/containers/consumerOnboarding/ConsumerOnboardingRoot.tsx
@@ -12,7 +12,6 @@ import {
 import DemoPrep from './DemoPrep';
 import OnboardingOverview from './OnboardingOverview';
 import WorkingWithOurAPIs from './WorkingWithOurAPIs';
-import RequestSandboxAccess from './RequestSandboxAccess';
 import RequestProductionAccess from './RequestProductionAccess';
 
 const ConsumerOnboardingRoot = (): JSX.Element => (
@@ -30,7 +29,6 @@ const ConsumerOnboardingRoot = (): JSX.Element => (
       <Switch>
         <Route exact path={CONSUMER_PATH} component={OnboardingOverview} />
         <Route exact path={CONSUMER_APIS_PATH} component={WorkingWithOurAPIs} />
-        <Route exact path={CONSUMER_SANDBOX_PATH} component={RequestSandboxAccess} />
         <Route path={CONSUMER_PROD_PATH} component={RequestProductionAccess} />
         <Route path={CONSUMER_DEMO_PATH} component={DemoPrep} />
       </Switch>
