--- src/containers/consumerOnboarding/components/sandbox/SandboxAccessFormLegacy.scss
@@ -1,7 +1,0 @@
-legend#typeAndApiLegend {
-  font-size: 1.6rem;
-  font-weight: 400;
-}
-.va-api-text-field {
-  width: 46rem;
-}
