+++ src/containers/documentation/Documentation.scss
@@ -20,6 +20,7 @@
   .swagger-ui {
     .wrapper {
       margin: unset;
+      max-width: 900px;
     }
 
     .information-container {
