+++ src/containers/documentation/components/sandbox/SandboxAccessSuccess.scss
