+++ src/containers/documentation/swaggerPlugins/VersionSelect.scss
@@ -34,6 +34,7 @@
   &.theme-light {
     background-color: #f1f1f1;
     margin: 1.6rem 0;
+    max-width: 900px;
     padding: 1.6rem 1rem;
   }
