+++ src/types/index.ts
@@ -9,9 +9,6 @@ declare global {
   }
 }
 
-export interface APIUrlFragment {
-  urlFragment: string;
-}
 export interface APIUrlSlug {
   urlSlug: string;
 }
